fix install script for MAC OS
On MAC OS, command line tool parameters should be placed right after the
command itself, and not the end. Such as, `cp /path/to/a /path/to/b -rf`
is invalid, while `cp -rf /path/to/a /path/to/b` is OK. While on Linux,
both the two patters are supported.